+++ DarkSoulsScripting/Injection/PSAPI.cs
@@ -9,8 +9,7 @@ namespace DarkSoulsScripting.Injection
         [DllImport("psapi.dll", SetLastError = true)]
         internal static extern bool EnumProcessModules(
             IntPtr hProcess,
-            [MarshalAs(UnmanagedType.LPArray, ArraySubType = UnmanagedType.U4)]
-            [In()] [Out()] uint[] lphModule,
+            [MarshalAs(UnmanagedType.LPArray, ArraySubType = UnmanagedType.U4)] [In()] [Out()] IntPtr[] lphModule,
             uint cb,
             [MarshalAs(UnmanagedType.U4)] ref uint lpcbNeeded);
